src/backend/requirements.txt
fastapi[standard]>=0.115.0
httpx>=0.28.1
openai>=1.58.0
sqlmodel>=0.0.22
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
